+++ crates/lib/src/games/fivem/mod.rs
@@ -1,0 +1,8 @@
+/// The implementation.
+/// Reference: [node-GameDig](https://github.com/gamedig/node-gamedig/blob/master/protocols/fivem.js)
+pub mod protocol;
+/// All types used by the implementation.
+pub mod types;
+
+pub use protocol::*;
+pub use types::*;